Fixes signature of list-queue-unfold-right
The signature of `list-queue-unfold-right` in the documentation was
missing the optional parameter `[ queue ]`, despite it being referenced
in the text below. I have added the optional parameter to the procedure
signature. Note that this changes nothing about the SRFI, as the
description of the procedure includes this parameter and the
implementation of the procedure has it as well. The signature was the
only thing that appeared to have missed this parameter.

NOTE: My editor also removed all extraneous whitespace at the end of
lines (equivalent to s/\s*$//g), so it looks like three lines were
edited instead of just the one.